refactor(pool-page): extract _tabs/swaps-tab
Move SwapsTab (~270 lines) out of page.tsx into _tabs/swaps-tab.tsx.
Tab queries POOL_SWAPS_COUNT, POOL_SWAPS_PAGE,
POOL_DAILY_SNAPSHOTS_CHART, and POOL_REBALANCES (the last one for
volume-chart annotations — it's a shared query, also fired by
RebalancesTab). The "// Tab content" header comment moves with the
extraction.

Behavior-preserving — verified via:
- pnpm test (1366/1366 green)
- pnpm typecheck (clean)
- chrome-devtools snapshot of /pool/<id>?tab=swaps vs baseline:
  zero structural changes, console clean

page.tsx: 1808 → 1540 lines (-268).
